requirements.txt
# Core dependencies
python>=3.9

# PDF processing
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# NLP and ML
spacy>=3.7.0
transformers>=4.35.0
sentence-transformers>=2.3.0
torch>=2.0.0

# LLM frameworks
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
openai>=1.6.0

# RAG and vector stores
chromadb>=0.4.18
faiss-cpu>=1.7.4
llama-index>=0.9.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Web interface
streamlit>=1.29.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
tqdm>=4.66.0

# Optional: for better PDF parsing
pymupdf>=1.23.0

# Optional: for XAI (if needed)
shap>=0.43.0

